Split the propose-check runner out of the gate module
propose_check_gate.py landed at 1411 lines, 89 under the 1500-line hard
cap, in the PR whose exhibit A is a file 84 lines over that cap. It also
holds two separable concerns: the decision (operator switches, which
checks to run, the verdict ledger, whether a propose is rejected) and the
execution (the sandboxed runner Job, its manifest, the verdict protocol).

Move the execution half to orchestrator/propose_check_runner.py. Gate is
765 lines, runner 704, both under the caps and each one seam. The gate
calls run_propose_checks through the module object so it stays a patch
seam for tests of the decision that never spawn a pod. No behaviour
change; a top-level .py either side, so the orchestrator image's
COPY orchestrator/*.py glob still ships both.
